Use Semantic Versioning for Git tags
With this commit regular expressions for dealing with Git tags
have been removed and replaced with a Semantic Versioning library.
This should improve maintainability there by offloading the version
voodoo as well as allow for future improvements such as warning for
major version bumps and more.

An imporant note however is that this removes support for version
tags which consist of just one number of in an X.Y format, as
Semantic Versioning mandates a complete X.Y.Z version tag at the
very least.